move L90Talents to /Talents
