+++ src/Parser/Druid/Balance/Modules/Features/Checklist.js
@@ -23,7 +23,7 @@ import StellarFlareUptime from './StellarFlareUptime';
 import MoonSpells from './MoonSpells';
 import LunarEmpowerment from './LunarEmpowerment';
 import SolarEmpowerment from './SolarEmpowerment';
-import L90Talents from './L90Talents';
+import L90Talents from '../Talents/L90Talents';
 
 import SoulOfTheArchdruid from '../../../Shared/Modules/Items/SoulOfTheArchdruid';
 
@@ -77,15 +77,15 @@ class Checklist extends CoreChecklist {
         const combatant = this.combatants.selected;
         return [
           new Requirement({
-            name: <Wrapper><SpellLink id={SPELLS.MOONFIRE_BEAR.id} icon/> uptime</Wrapper>,
+            name: <Wrapper><SpellLink id={SPELLS.MOONFIRE_BEAR.id} /> uptime</Wrapper>,
             check: () => this.moonfireUptime.suggestionThresholds,
           }),
           new Requirement({
-            name: <Wrapper><SpellLink id={SPELLS.SUNFIRE.id} icon/> uptime</Wrapper>,
+            name: <Wrapper><SpellLink id={SPELLS.SUNFIRE.id} /> uptime</Wrapper>,
             check: () => this.sunfireUptime.suggestionThresholds,
           }),
           new Requirement({
-            name: <Wrapper><SpellLink id={SPELLS.STELLAR_FLARE_TALENT.id} icon/> uptime</Wrapper>,
+            name: <Wrapper><SpellLink id={SPELLS.STELLAR_FLARE_TALENT.id}  /> uptime</Wrapper>,
             check: () => this.stellarFlareUptime.suggestionThresholds,
             when: combatant.hasTalent(SPELLS.STELLAR_FLARE_TALENT.id),
           }),
@@ -164,7 +164,7 @@ class Checklist extends CoreChecklist {
             check: () => this.l90Talents.suggestionThresholds,
           }),
           new Requirement({
-            name: <Wrapper>Picked the right talent with <ItemLink id={ITEMS.SOUL_OF_THE_ARCHDRUID.id} icon/></Wrapper>,
+            name: <Wrapper>Picked the right talent with <ItemLink id={ITEMS.SOUL_OF_THE_ARCHDRUID.id} /></Wrapper>,
             check: () => this.soulOfTheArchdruid.suggestionThresholds,
             when: this.soulOfTheArchdruid.active,
           }),
